lazyビュー読み込み時のホワイトアウトを修正
ガチャ・まち編集などのlazyビューへ遷移して「よみこみ中」になると、
画面が白いまま固まることがあった。

原因: SuspenseがAnimatePresence(mode=wait)の外側にあったため、
チャンク読み込みでサスペンドするとAnimatePresenceのツリーごと
非表示化され、退場アニメーションの完了追跡が壊れていた。
その結果、
- 前の画面(ホーム)がopacity:0のまま残り、画面が実質白紙になる
- 読み込み中のタップ(設定・ホーム等)がすべて無視される
- 通信が遅い/失敗する端末では白画面のまま復帰しない

修正: Suspense境界をPageWrapper/FullScreenWrapperのmotion.div内側へ
移動。サスペンドがアニメーション対象の外へ漏れなくなり、
- 読み込み中もスピナーが画面内に表示され、他ボタンも即座に反応
- チャンク取得失敗時はErrorBoundaryの再試行画面が表示される
ことをPlaywrightで遅延・404・読み込み中の割り込み操作を含めて確認。
